Make the active filter not just based on status but on the actual download/upload state
